Rename locale folder for better compatibility with i18n-ally
