+++ ui/v2.5/src/locales/README.md
