+++ ui/v2.5/src/locales/en-US.json
@@ -2,5 +2,6 @@
   "eye_color": "Eye Color",
   "favourite": "Favorite",
   "hair_color": "Hair Color",
-  "organized": "Organized"
+  "organized": "Organized",
+  "path": "Path"
 }
