+++ ui/v2.5/src/locales/index.ts
